Move common linux detection functions to separate file. Move WSL detection functions to separate file. Update projet files for Windows, linux and MacOS
