+++ client/Makefile.curl.am
@@ -43,6 +43,7 @@ boinc_client_SOURCES = \
     gui_http.cpp \
     gui_rpc_server.cpp \
     gui_rpc_server_ops.cpp \
+    hostinfo_linux.cpp \
     hostinfo_network.cpp \
     hostinfo_unix.cpp \
     http_curl.cpp \
