+++ win_build/boinc_cli_vs2013.vcxproj
@@ -354,6 +354,8 @@
     <ClCompile Include="..\client\app_start.cpp" />
     <ClCompile Include="..\client\async_file.cpp" />
     <ClCompile Include="..\client\coproc_sched.cpp" />
+    <ClCompile Include="..\client\hostinfo_linux.cpp" />
+    <ClCompile Include="..\client\hostinfo_wsl.cpp" />
     <ClCompile Include="..\client\mac_address.cpp" />
     <ClCompile Include="..\client\project_list.cpp" />
     <ClCompile Include="..\client\thread.cpp" />
